client side doing the thing
