+++ web/src/i18n/locales/en.yaml
@@ -46,6 +46,13 @@
             city: City
             state: State
             zip: ZIP
+      pointsOfContact:
+        title: Points of Contact
+        labels:
+          name: Name
+          position: Position
+          email: Email address
+          addButton: Add point-of-contact
     overview: 
       title: "Program Overview"
       helpText: "Provide a high level summary of your state's program, history, and future vision."
